+++ src/components/CardNameInput.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 
 
-export function CardName() {
+export function CardNameInput() {
   const [cardName, setCardName] = useState("");
